[RL] Fix flashmask reward training
