--- llm/alignment/rm/models
@@ -1,1 +1,0 @@
-../ppo/models
