+++ llm/config/llama/rm_flashmask_argument.json
@@ -27,7 +27,7 @@
     "sharding_parallel_degree": 1,
     "sharding": "stage1",
     "use_flash_attention": true,
-    "flash_mask":false,
+    "flash_mask":true,
     "recompute": false,
     "recompute_granularity": "full",
     "benchmark": false,
